Updates for multiple LOS versions

+++ source/replace-captive-portal-detection-url.sh
@@ -1,27 +1,35 @@
 #!/bin/bash
 
-# Replace the default captive portal detection URL in the TV app.
-cd ~/android/lineage/packages/apps/TV/src/com/android/tv/util
+# Update for all versions of LOS that we have.
+for LOSPATHNAME in ~/android/lineage-*; do
+        LOSDIRNAME=$(basename $LOSPATHNAME)
 
-sed 's/clients3.google.com/wunderment.org/' NetworkUtils.java > NetworkUtils.java.new 
+	# Replace the default captive portal detection URL in the TV app.
+	cd ~/android/$LOSDIRNAME/packages/apps/TV/src/com/android/tv/util
 
-rm NetworkUtils.java
-mv NetworkUtils.java.new NetworkUtils.java
+	sed 's/clients3.google.com/wunderment.org/' NetworkUtils.java > NetworkUtils.java.new 
 
-# Replace the captive portal detection URL in the setup wizard.
-cd ~/android/lineage/packages/apps/SetupWizard/src/org/lineageos/setupwizard
+	rm NetworkUtils.java
+	mv NetworkUtils.java.new NetworkUtils.java
 
-sed 's/clients3.google.com/wunderment.org/' CaptivePortalSetupActivity.java > CaptivePortalSetupActivity.java.new 
+	# Replace the captive portal detection URL in the setup wizard.
+	cd ~/android/$LOSDIRNAME/packages/apps/SetupWizard/src/org/lineageos/setupwizard
 
-rm CaptivePortalSetupActivity.java
-mv CaptivePortalSetupActivity.java.new CaptivePortalSetupActivity.java
+	sed 's/clients3.google.com/wunderment.org/' CaptivePortalSetupActivity.java > CaptivePortalSetupActivity.java.new 
 
-# Replace the default captive portal deteciton in the network monitor.
-cd ~/android/lineage/frameworks/base/services/core/java/com/android/server/connectivity
+	rm CaptivePortalSetupActivity.java
+	mv CaptivePortalSetupActivity.java.new CaptivePortalSetupActivity.java
 
-sed 's/connectivitycheck.gstatic.com/wunderment.org/' NetworkMonitor.java > NetworkMonitor.java.new 
-sed 's/www.google.com/wunderment.org/' NetworkMonitor.java.new > NetworkMonitor.java 
-sed 's/gen_204/generate_204/' NetworkMonitor.java > NetworkMonitor.java.new 
-sed 's/play.googleapis.com/wunderment.org/' NetworkMonitor.java.new > NetworkMonitor.java 
+	# Replace the default captive portal deteciton in the network monitor.
+	cd ~/android/$LOSDIRNAME/frameworks/base/services/core/java/com/android/server/connectivity
 
-rm NetworkMonitor.java.new
+	if [ -f "NetworkMonitor.java" ];
+	then
+		sed 's/connectivitycheck.gstatic.com/wunderment.org/' NetworkMonitor.java > NetworkMonitor.java.new 
+		sed 's/www.google.com/wunderment.org/' NetworkMonitor.java.new > NetworkMonitor.java 
+		sed 's/gen_204/generate_204/' NetworkMonitor.java > NetworkMonitor.java.new 
+		sed 's/play.googleapis.com/wunderment.org/' NetworkMonitor.java.new > NetworkMonitor.java 
+
+		rm NetworkMonitor.java.new
+	fi
+done
